requirements.txt
# LLM
anthropic>=0.96.0

# NetEase Music
pyncm>=1.5.0
qrcode[pil]>=7.4.0

# Music metadata / lyrics
requests>=2.31.0
beautifulsoup4>=4.12.0

# Spotify / Last.fm / Weather APIs
# (also use requests above)

# UI
PyQt6>=6.6.0

# TTS & Speech
edge-tts>=7.0.0
SpeechRecognition>=3.10.0
pyaudio>=0.2.14

# Config
python-dotenv>=1.0.0
